eshell: Cleanup use-package definition
This is a subjective cleanup to make things easier to read, and also
remove the need for loading an entire function definition when
entering eshell-mode, instead creating a defalias when entering
eshell-mode. I'm far less embarrassed with this use-package definition
now though there is plenty of weirdness in other sections to resolve.

Something I'll be looking to do in the near future is localizing the
changes made in baemacs so that it's more obvious when a non-standard
function is being used.