check for any stale configuration files on start
Prior to this commit, we only checked if config.el was stale. This
lead to situations where early-init.el or init.el could be older than
their respective byte-compiled .elc files. As of this commit we now
check if any of these files are stale when compared to README.org
which resolves this corner-case.